+++ src/layouts/TutorDashboardLayout/TopBar.js
@@ -39,6 +39,7 @@ import AccountCircleIcon from '@material-ui/icons/AccountCircle';
 import InsertInvitationIcon from '@material-ui/icons/InsertInvitation';
 import AccountBalanceWalletIcon from '@material-ui/icons/AccountBalanceWallet';
 import Notifications from "react-notifications-menu";
+import { useConfirm } from 'material-ui-confirm';
 import 'src/React-Notifs.css'
 
 import 'intro.js/introjs.css';
@@ -60,9 +61,15 @@ const TopBar = ({
   const classes = useStyles();
   const anchor = 'left';
   const [notifications] = useState([]);
+  const confirm = useConfirm();
   const logout = () => {
-    localStorage.clear()
-    window.location.replace('/')
+    confirm({ title:'Logout' , description: 'Would you like to logout?' })
+      .then(() => {
+        localStorage.clear()
+        window.location.replace('/')
+      })
+      .catch(() => {
+      });
   }
 
   const [state, setState] = React.useState({
